pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-account-scraper"
version = "0.1.0"
description = "Export Telegram account chats to Markdown files."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115,<1",
  "hdbscan>=0.8,<1",
  "numpy>=2,<3",
  "openai>=1.78,<2",
  "pinecone>=6,<7",
  "psycopg[binary]>=3.2,<4",
  "redis>=5.2,<6",
  "scikit-learn>=1.6,<2",
  "telethon>=1.40,<2",
  "typer>=0.12,<1",
  "umap-learn>=0.5,<1",
  "uvicorn>=0.34,<1",
]

[dependency-groups]
dev = [
  "httpx>=0.28,<1",
  "pytest>=8,<9",
]

[project.scripts]
telegram-scraper = "telegram_scraper.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
